Use Diesel to create a table on start-up (#17)
dhb_postgres_heroku uses a pure-Rust Postgres client, which does not
implement the same traits that Diesel is expecting:

    error[E0277]: the trait bound `dhb_postgres_heroku::Client: Connection` is not satisfied
       --> src/main.rs:162:30
        |
    69  | embed_migrations!("migrations");
        | -------------------------------- required by this bound in `embedded_migrations::run`
    ...
    162 |     embedded_migrations::run(&db_client);
        |                              ^^^^^^^^^^ the trait `Connection` is not implemented for `dhb_postgres_heroku::Client`
        |
        = note: required because of the requirements on the impl of `diesel_migrations::MigrationConnection` for `dhb_postgres_heroku::Client`

so that has been replaced with the default mechanism of connecting to Postgres
from Diesel.  The default postgres behavior is not to check validity of SSL
certificates, so this should work with Heroku's self-signed stuff.

In addition, the global Serenity client's `data` is shared across all of
the threads and futures, so pull in Diesel's r2d2 support to share a
`ConnectionManager` rather than trying to share a single `Connection`.